@@ -16,8 +16,8 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Sequence
 
-if sys.version_info < (3, 7):
-    sys.exit("Python 3.7 or above is required to install PDM.")
+if sys.version_info < (3, 8):
+    sys.exit("Python 3.8 or above is required to install PDM.")
 
 _plat = platform.system()
 MACOS = _plat == "Darwin"
